rename trainer to feature generation
