+++ feature_generation/datasets/EMIP.py
@@ -1,9 +1,9 @@
-# from trainer.utils import convert_categorical_labels_to_numerical
-from trainer.Labels import Labels
+# from feature_generation.utils import convert_categorical_labels_to_numerical
+from feature_generation.Labels import Labels
 import pandas as pd
 from itertools import takewhile
 import time
-from trainer.datasets.Timeseries import Timeseries
+from feature_generation.datasets.Timeseries import Timeseries
 
 
 class EMIP(Timeseries):
