+++ scripts/generate_features.sh
@@ -33,7 +33,7 @@ COMMIT_MESSAGE_WITHOUT_BLANK="${COMMIT_MESSAGE//[[:blank:]]/_}"
 COMMIT_MESSAGE_WITHOUT_NEWLINE="${COMMIT_MESSAGE_WITHOUT_BLANK//$'\n'/_}"
 JOB_NAME="${RUN_TYPE}_${NOW}_${COMMIT_MESSAGE_WITHOUT_NEWLINE////_}_${COMMIT_HASH}"
 JOB_DIR="gs://$BUCKET_ID/models/$JOB_NAME"
-PACKAGE_PATH=trainer
+PACKAGE_PATH=feature_generation
 MAIN_TRAINER_MODULE=$PACKAGE_PATH.task
 REGION=europe-west1
 CONFIG_FILE=config/config.yaml
